# api-specs/swagger.json
{
  "info": {
    "title": "Application APIs",
    "version": "1.0.0",
    "description": "Application APIs"
  },
  "basePath": "/",
  "swagger": "2.0",
  "paths": {
...
    "/subject": {
      "get": {
        "description": "get Subject",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "get Subject",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Subjectresponse"
                  }
                },
                "pageInfo": {
                  "type": "object",
                  "properties": {
                    "pageSize": {
                      "type": "integer"
                    },
                    "pageLimit": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "save Subject",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Subject",
            "schema": {
              "type": "object",
              "properties": {
                "Subject": {
                  "$ref": "#/definitions/Subject"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "save Subject",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/Subjectresponse"
                }
              }
            }
          }
        }
      }
    },
    "/subject/{id}": {
      "get": {
        "description": "get Subject with id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Subject ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "get Subject",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/Subjectresponse"
                }
              }
            }
          },
          "404": {
            "description": "Subject not found",
            "content": {}
          }
        }
      },
      "delete": {
        "description": "remove Subject with id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Subject ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "delete Subject",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/Subjectresponse"
                }
              }
            }
          },
          "404": {
            "description": "Subject not found",
            "content": {}
          }
        }
      },
      "put": {
        "description": "update Subject with id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Subject ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "Subject",
            "schema": {
              "type": "object",
              "properties": {
                "Subject": {
                  "$ref": "#/definitions/Subject"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "update Subject",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/Subjectresponse"
                }
              }
            }
          },
          "404": {
            "description": "Subject not found",
            "content": {}
          }
        }
      }
    },
    "/teacher": {
      "get": {
        "description": "get Teacher",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "get Teacher",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Teacherresponse"
                  }
                },
                "pageInfo": {
                  "type": "object",
                  "properties": {
                    "pageSize": {
                      "type": "integer"
                    },
                    "pageLimit": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "save Teacher",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Teacher",
            "schema": {
              "type": "object",
              "properties": {
                "Teacher": {
                  "$ref": "#/definitions/Teacher"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "save Teacher",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/Teacherresponse"
                }
              }
            }
          }
        }
      }
    },
    "/teacher/{id}": {
      "get": {
        "description": "get Teacher with id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Teacher ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "get Teacher",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/Teacherresponse"
                }
              }
            }
          },
          "404": {
            "description": "Teacher not found",
            "content": {}
          }
        }
      },
      "delete": {
        "description": "remove Teacher with id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Teacher ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "delete Teacher",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/Teacherresponse"
                }
              }
            }
          },
          "404": {
            "description": "Teacher not found",
            "content": {}
          }
        }
      },
      "put": {
        "description": "update Teacher with id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Teacher ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "Teacher",
            "schema": {
              "type": "object",
              "properties": {
                "Teacher": {
                  "$ref": "#/definitions/Teacher"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "update Teacher",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/Teacherresponse"
                }
              }
            }
          },
          "404": {
            "description": "Teacher not found",
            "content": {}
          }
        }
      }
    },
    "/teachers-subject": {
      "get": {
        "description": "get TeachersSubject",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "get TeachersSubject",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/TeachersSubjectresponse"
                  }
                },
                "pageInfo": {
                  "type": "object",
                  "properties": {
                    "pageSize": {
                      "type": "integer"
                    },
                    "pageLimit": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "save TeachersSubject",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "TeachersSubject",
            "schema": {
              "type": "object",
              "properties": {
                "TeachersSubject": {
                  "$ref": "#/definitions/TeachersSubject"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "save TeachersSubject",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/TeachersSubjectresponse"
                }
              }
            }
          }
        }
      }
    },
    "/teachers-subject/{id}": {
      "get": {
        "description": "get TeachersSubject with id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "TeachersSubject ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "get TeachersSubject",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/TeachersSubjectresponse"
                }
              }
            }
          },
          "404": {
            "description": "TeachersSubject not found",
            "content": {}
          }
        }
      },
      "delete": {
        "description": "remove TeachersSubject with id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "TeachersSubject ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "delete TeachersSubject",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/TeachersSubjectresponse"
                }
              }
            }
          },
          "404": {
            "description": "TeachersSubject not found",
            "content": {}
          }
        }
      },
      "put": {
        "description": "update TeachersSubject with id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "TeachersSubject ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "TeachersSubject",
            "schema": {
              "type": "object",
              "properties": {
                "TeachersSubject": {
                  "$ref": "#/definitions/TeachersSubject"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "update TeachersSubject",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/TeachersSubjectresponse"
                }
              }
            }
          },
          "404": {
            "description": "TeachersSubject not found",
            "content": {}
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Subject": {
      "properties": {
        "name": {
          "type": "string"
        },
        "totalHours": {
          "type": "integer"
        },
        "courseId": {
          "type": "integer"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "updatedAt": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "Subjectresponse": {
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "totalHours": {
          "type": "integer"
        },
        "courseId": {
          "type": "integer"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "updatedAt": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "Teacher": {
      "properties": {
        "workingTypeId": {
          "type": "integer"
        },
        "personInfoId": {
          "type": "integer"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "updatedAt": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "Teacherresponse": {
      "properties": {
        "id": {
          "type": "integer"
        },
        "workingTypeId": {
          "type": "integer"
        },
        "personInfoId": {
          "type": "integer"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "updatedAt": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "TeachersSubject": {
      "properties": {
        "subjectId": {
          "type": "integer"
        },
        "teacherId": {
          "type": "integer"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "updatedAt": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "TeachersSubjectresponse": {
      "properties": {
        "id": {
          "type": "integer"
        },
        "subjectId": {
          "type": "integer"
        },
        "teacherId": {
          "type": "integer"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "updatedAt": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
...
  },
...
  "securityDefinitions": {},
...
}